events: move some helpers to events mod
