+++ src/events/mod.rs
@@ -19,6 +19,7 @@ pub(crate) mod display;
 pub(crate) use display::*;
 
 pub(crate) mod file;
+pub(crate) mod net;
 
 pub(crate) mod common;
 pub(crate) use common::*;
